backend/requirements.txt
aniso8601==8.0.0
attrs==19.3.0
boto3==1.9.253
botocore==1.12.253
certifi==2019.9.11
chardet==3.0.4
Click==7.0
docutils==0.15.2
Flask==1.1.1
Flask-Cors==3.0.8
flask-restplus==0.13.0
h11==0.8.1
httptools==0.0.13
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
jsonschema==3.1.1
MarkupSafe==1.1.1
more-itertools==7.2.0
numpy==1.17.3
pandas==0.25.2
Pillow==6.2.1
pkg-resources==0.0.0
pyrsistent==0.15.4
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.6
uvicorn==0.9.1
uvloop==0.13.0
websockets==8.0.2
Werkzeug==0.16.0
zipp==0.6.0
